Create step2_exactROIandmeasure_ImageJ
This code happens after segmentation with cellpose. In the previous step, png images have been generated, so this step will extract them in ImageJ and save them as ROI files in the same folder. In the meantime, open the tiff file with the same name (first part of the name) and make measurement, then save each measurement table as csv with the same file name